+++ CHANGELOG.md
@@ -16,6 +16,8 @@ created.
 - ([#465](https://github.com/microsoft/InnerEye-DeepLearning/pull/465/)) Adding ability to run segmentation inference
 module on test data with partial ground truth files. (Also [522](https://github.com/microsoft/InnerEye-DeepLearning/pull/522).)
 - ([#502](https://github.com/microsoft/InnerEye-DeepLearning/pull/502)) More flags for fine control of when to run inference.
+- ([#537](https://github.com/microsoft/InnerEye-DeepLearning/pull/537)) Print a list of installed Python packages at the start of the 
+runner, to enable better diagnostics.
 - ([#492](https://github.com/microsoft/InnerEye-DeepLearning/pull/492)) Adding capability for regression tests for test
 jobs that run in AzureML.